third_party/blink/web_tests/virtual/document-isolation-policy/external/wpt/html/document-isolation-policy/shared-workers.tentative.https-expected.txt

This is a testharness.js-based test.
[FAIL] isolate-and-require-corp (derived from blob URL creator)
  assert_equals: expected "failure" but got "success"
Harness: the test ran to completion.
